# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Threat Modeling Agent API
  version: 0.1.0
  description: |
    REST API for the Threat Modeling Agent SaaS platform.

    ## Authentication
    All endpoints (except the auth callback) require a valid Bearer token issued by WorkOS.
    Tokens are validated server-side on every request. Client-side validation is not sufficient.

    ## Tenant scoping
    All data is scoped to an organisation. The `org_id` is extracted from the validated JWT —
    it is never accepted from request parameters or body.

    ## Rate limiting
    Authentication and job submission endpoints are rate-limited.
    See individual endpoint descriptions for details.

    ## Security
    This API is governed by CLAUDE.md and docs/specs/06-security.md.
    All responses to authenticated requests include `Cache-Control: no-store`.
servers:
  - url: https://api.tmsa.example.com/v1
    description: Production (West Europe)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: WorkOS-issued JWT. Validated server-side on every request.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organisation UUID. Must match the org_id in the authenticated user's JWT.
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response.
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code. Never contains stack traces or internal paths.
        message:
          type: string
          description: Human-readable summary. Never contains internal details.
        correlationId:
          type: string
          format: uuid

    # ── Pagination ───────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
        totalCount:
          type: integer

    # ── Auth ─────────────────────────────────────────────────────────────────────
...
    JobStatus:
      type: string
      enum:
        - PENDING
        - PARSING
        - NORMALIZING
        - AWAITING_REVIEW
        - CLASSIFYING
        - ANALYZING
        - SYNTHESIZING
        - COMPLETE
        - FAILED
        - PARTIAL
    JobSummary:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        artifactType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    JobDetail:
      allOf:
        - $ref: '#/components/schemas/JobSummary'
        - type: object
          properties:
            errorCode:
              type: string
              nullable: true
            threatCount:
              type: integer
              nullable: true
            confirmedThreatCount:
              type: integer
              nullable: true

    # ── Architecture ─────────────────────────────────────────────────────────────
...
paths:
...
  /orgs/{orgId}/jobs:
    parameters:
      - $ref: '#/components/parameters/OrgId'

    get:
      operationId: listJobs
      summary: List analysis jobs for the organisation
      tags: [jobs]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      operationId: submitJob
      summary: Submit a new threat modeling analysis
      description: |
        Accepts a multipart/form-data body with the architecture artifact.
        Supported artifact types: image (PNG, JPG, GIF, WebP), PlantUML (.puml),
        Mermaid (.md, .mmd), Draw.io XML (.drawio, .xml), plain text (.txt).
        Maximum file size: 10 MB.
        Rate-limited: 10 job submissions per hour per organisation.
      tags: [jobs]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [artifact]
              properties:
                artifact:
                  type: string
                  format: binary
                  description: The architecture diagram or description file.
                title:
                  type: string
                  maxLength: 255
                  description: Optional human-readable title for this analysis.
      responses:
        '202':
          description: Job accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetail'
        '413':
          description: Artifact too large (max 10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: Unsupported artifact type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}/jobs/manual:
    parameters:
      - $ref: '#/components/parameters/OrgId'

    post:
      operationId: createManualJob
      summary: Create a manual job with no file upload
      description: |
        Creates a job in AWAITING_REVIEW status with an empty architecture model.
        The user builds the architecture by adding elements via POST /elements.
        Once ready, confirm via POST /architecture/confirm to trigger analysis.
        Rate-limited: same tier as job submission.
      tags: [jobs]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 255
                systemPurpose:
                  type: string
                  maxLength: 5000
                  description: Brief description of what the system does.
      responses:
        '201':
          description: Manual job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetail'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}/jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/JobId'

    get:
      operationId: getJob
      summary: Get job status and summary
      tags: [jobs]
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetail'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteJob
      summary: Delete a job and all its associated data
      tags: [jobs]
      responses:
        '204':
          description: Job deleted
        '409':
          description: Job is in progress and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Architecture (canonical model) ───────────────────────────────────────────
...